Since C passes arguments to functions by value, there is no direct way for the called function to alter a variable in the calling function. Since the operator & produces the address of a variable, &a is a pointer to a.
When necessary, it is possible to arrange for a function to modify a variable in
a calling routine. The caller must provide the address of the variable to be set
(technically a pointer to the variable), and the called function must declare
the parameter to be a pointer and access the variable indirectly through it.

A pointer to a pointer is a form of multiple indirection, or a chain of pointers.
Normally, a pointer contains the address of a variable. When we define a pointer
to a pointer, the first pointer contains the address of the second pointer,
which points to the location that contains the actual value.